Menyederhanakan MainViewModel
