--- app/src/main/java/com/humam/smarthydro/data/DataDao.kt
@@ -1,8 +1,0 @@
-package com.humam.smarthydro.data
-
-import com.humam.smarthydro.model.WaterLevel
-import kotlinx.coroutines.flow.Flow
-
-interface DataDao {
-    suspend fun getWaterLevels(): Flow<List<WaterLevel>>
-}
