+++ app/src/main/java/com/humam/smarthydro/viewmodels/MainViewModel.kt
@@ -1,23 +1,57 @@
 package com.humam.smarthydro.viewmodels
 
 import androidx.lifecycle.ViewModel
-import com.humam.smarthydro.data.DataDB
-import com.humam.smarthydro.data.DataDao
+import androidx.lifecycle.viewModelScope
+import com.google.firebase.database.DataSnapshot
+import com.google.firebase.database.DatabaseError
+import com.google.firebase.database.FirebaseDatabase
+import com.google.firebase.database.ValueEventListener
+import com.google.firebase.database.getValue
 import com.humam.smarthydro.model.WaterLevel
-import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.channels.awaitClose
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.callbackFlow
+import kotlinx.coroutines.flow.stateIn
 
 class MainViewModel : ViewModel() {
 
-    private val dataDao: DataDao = DataDB.getInstance().dao
+    companion object {
+        const val WATER_LEVELS = "water-level"
+    }
 
-    private val _waterLevels: MutableStateFlow<List<WaterLevel>> = MutableStateFlow(emptyList())
-    val waterLevels: StateFlow<List<WaterLevel>> = _waterLevels.asStateFlow()
+    private val db = FirebaseDatabase.getInstance()
 
-    suspend fun getWaterLevels() {
-        dataDao.getWaterLevels().collect {
-            _waterLevels.value = it
+    private val _waterLevels: Flow<List<WaterLevel>> = callbackFlow {
+        val dataRef = db.getReference(WATER_LEVELS)
+        val listener = dataRef.addValueEventListener(object : ValueEventListener {
+            override fun onDataChange(snapshot: DataSnapshot) {
+                trySend(getWaterLevels(snapshot))
+            }
+
+            override fun onCancelled(error: DatabaseError) {
+                close(error.toException())
+            }
+        })
+
+        awaitClose { dataRef.removeEventListener(listener) }
+    }
+
+    val waterLevels: StateFlow<List<WaterLevel>> = _waterLevels.stateIn(
+        scope = viewModelScope,
+        started = SharingStarted.WhileSubscribed(),
+        initialValue = emptyList()
+    )
+
+    private fun getWaterLevels(snapshot: DataSnapshot): List<WaterLevel> {
+        val waterLevels = mutableListOf<WaterLevel>()
+        snapshot.children.forEach { data ->
+            val waterLevel = data.getValue<WaterLevel>()
+            if (waterLevel != null) {
+                waterLevels.add(waterLevel)
+            }
         }
+        return waterLevels
     }
 }
